use empty auth provider
