+++ tests/App.test.tsx
@@ -1,9 +1,14 @@
 import { render, screen } from "@testing-library/react";
 import App from "../src/App";
+import { AuthProvider } from "react-oidc-context";
 
 describe("App", () => {
   it("should render landing page when no user session active", () => {
-    render(<App />);
+    render(
+      <AuthProvider>
+        <App />
+      </AuthProvider>
+    );
 
     expect(screen.getByText("Login With SAMS")).toBeInTheDocument();
   });
